requirements.txt
python-dotenv
requests
opencv-python